// src/stores/slices/Items/type.ts
export interface ItemType {
    id: number;
    title: string;
    subtitle: string;
    description: string;
    about: string;
    image: string;
    star: number;
    value: number;
    inSale: boolean;
    percentage: number;
    isNew: boolean;
    sku: string;
    category: string;
    tags: string[];
    quantity: number;
  }